backend/requirements.txt
# Multi-Agent Content Generation Pipeline Requirements

# ============ Core Framework ============
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# ============ LLM Providers ============
openai>=1.50.0

# ============ Web Framework ============
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ============ Data & Validation ============
pydantic>=2.9.0
python-multipart>=0.0.12

# ============ Utilities ============
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.5.0

# ============ Development ============
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.7.0
mypy>=1.12.0

# ============ Optional: Monitoring ============
# prometheus-client>=0.21.0
# opentelemetry-api>=1.27.0
